Store htpasswd files in Secrets instead of ConfigMaps
The htpasswd files for Ironic and Inspector contains clear text
usernames and hashed passwords so it is better to store them in Secrets.

Depending on how exactly Ironic is deployed this could be a breaking
change that requires manual action from the user.
I have tested this with the
[deploy.sh](https://github.com/metal3-io/baremetal-operator/blob/main/tools/deploy.sh)
script and confirmed that it is working. Re-deploying Ironic, with the
updated kustomization using the script, automatically creates the new
Secrets and configures Ironic and Inspector to use them instead of the
ConfigMaps.

Note that the ConfigMaps are **not** automatically removed. Ideally, the
user should remove the ConfigMaps and change the credentials. The
ConfigMaps in question are named
`baremetal-operator-ironic-htpasswd-<random-hash>` and
`baremetal-operator-ironic-inspector-htpasswd-<random-hash>` and are
located in the `baremetal-operator-system` Namespace by default.

Note that if the credentials are changed, they must also be updated for
BMO. This can be done in the same way by re-deploying using the script.