Use same settings for provisioner and controller logs
Previously the ironic provisioner log was created in a separate module,
and didn't have access to the CLI arguments. Originally the provisioner
was always in structured mode even when the -dev flag was passed to put
it into development mode, and then after #1175 vice-versa.

Create the provisioning logger in main.go using the same settings as the
controller logger, and pass it in to the provisioner factory.

Fixes #1248