Validate diskPath in rootDeviceHints
In the webhook, check for diskPaths given in the root device hints that
cannot possibly be matched.

In a new API, this could be validated with a regex pattern in the CRD
itself, but that would be an incompatible API change so it will have to
wait for an API version bump.