Move hardware profile code to the api module
Since this code now defines how we map one API concept (the hardware
profile string) to default values for other API concepts (root device
hints), it is effectively part of the API definition. It is also not
expected to exist in future API versions. The webhook for translating
hosts between versions will need to access this code, so it will need to
be inside the API module. Finally, similar to the hardwareutils module,
this is useful in other contexts and moving it out of the main module
eliminates extraneous external dependencies being dragged with it.