Stop linting when building the image
When we build the image, the code has already been through code review,
and CI. Everything has been generated, and linted, and addressed. At the
build stage, we just want to compile the binary.

Furthermore, the golangci-lint program seems to require attention every
time we upgrade Go in the builder. This breaks the build, and we have to
do emergency upgrades to get it unstuck.

Lastly, we use the Makefile target `manager` instead of the low-level Go
command.  This gives us build-time information in the binary.