Add e2e tests
This adds a test suite for e2e tests. So far there is only a few tests
related to inspection, but it should be enough to validate the approach.

My hope is that this framework will be flexible enough to allow tests
both in CI, using VMs, and on actual hardware. The tests can be
configured through a config file. They can run on an existing cluster or
set up a temporary kind cluster automatically. Similarly, BMO, Ironic
and Cert-manager can all be deployed or not depending on configuration.

Logs are collected for BMO and Ironic when deployed automatically, even
in existing clusters.