Conditionally pass preprovisioningNetworkData to Ironic
When no PreprovisioningImage integration is enabled we have no way
currently to retrieve the preprovisioningNetworkData in the IPA ramdisk
which prevents the Ironic DHCP-less flow[1] from being possible.

With this change we'll pass the preprovisioningNetworkData to
Ironic so it can be injected into the ramdisk image without
requiring any custom image-build controller.

[1] https://docs.openstack.org/ironic/latest/admin/dhcp-less.html