Add e2e external inspection test
The test simply creates a BMH with annotations for disabling inspection
and for adding hardware details. Then it checks that the BMH becomes
available without going through inspection.

This also adds cert-manager images to the e2e config so that if they
happen to be available locally, then they can be loaded into the kind
cluster (if used). It helps save some time when running the test
repetedly.

Minor details:
- Don't exit on e2e test failure, gather artifacts/logs first
- Destroy the network before trying to undefine it
- Wait for namespace deletion between e2e tests since we are using the
  same VM for the BMHs in multiple tests