Use kustomize components for BMO
This restructures the BMO kustomizations to make use of components
(similar to what was already done for the Ironic kustomizations).

It also makes the kustomizations usable "off the shelf" by removing the
secret generation. Previously it was necessary to clone the repository
and add the relevant files (e.g. by running `tools/deploy.sh`) before it
was possible to build the kustomizations that included secrets generated
from the files. With this commit, the secrets will NOT be generated as
part of the kustomizations and they can thus be used as is. However, the
user is instead required to add the necessary secrets.

The secrets can be added in any way. One common way is that the users
creates their own overlays that add the secrets. These kustomizations
can be based on and use the kustomizations in the repo since they are
now buildable without changes. Another way to add the secrets is via the
external secrets operator.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>