E2E: Use real ISO for live-ISO test
This makes the live-ISO test use a separate image so we can use an
actual ISO. Also rename the BMCs config files to have the protocol
instead of the emulator in the name and added a third file for
redfish-virtualmedia.

The live-ISO test will be skipped if booting with (i)PXE since it would
require a special image and configuration to verify.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>