bump gophercloud/v2 to beta.3 + some commits to fix go.mod
When bumping gophercloud/v2 to v2.0.0-beta.2, it brought along a
patch version to go directive in root go.mod. This set up a very strict
golang requirement on BMO, which was made worse by the golang bump to
v1.21.9.

My PR in gophercloud fixed this issue, and we need to take in this
v2.0.0-beta.3 + some commtis to include the fix, in order to not release
BMO 0.6.0 with this requirement.

In main, we should aim to bump to beta.4 as soon as it is released and
preferably cherry-pick that to release-0.6 branch too.

Signed-off-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>