Fix Status logic for HostFirmwareComponents
Make Updates and Components optional in HostFirmwareComponentsStatus,
so we don't get errors when they are empty.

The UpdatedAt field for FirmwareComponentStatus can be nil, this
caused runtime error in the BMO.

Change logic in GetFirmwareComponents, so we fail earlier in case
is not supported, now we raise a specific error
ErrFirmwareUpdateUnsupported.

Simplify the logic in the Controller, so we handle conditions in only
one function

Signed-off-by: Iury Gregory Melo Ferreira <imelofer@redhat.com>