Workflow pipeline with unit tests
This adds a unit test workflow and restructures the github workflows
so that we run golangci-lint and unit tests first. Only when they have
succeded will the pipeline continue with e2e tests.
The goal is to conserve resources and avoid running pointless jobs.
The commit also removes unnecessary permissions and avoids running the
jobs if the PR is a draft.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>