Migrate information away from docs/api.md
The motivation of this change is to avoid duplicating information
between the inline API docs, the user guide and this page.

Some useful bits of information have been migrated to the inline docs,
which I also proof-read for mistakes. Then most of the content of api.md
has been replaced with references to the rendered CR docs and the user
guide.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>