Halt BMH transition on invalid BIOS setting
This update ensures that the BMH (Bare Metal Host) does not transition
to the next state when an invalid BIOS setting is detected. This pause
allows the user to correct the settings before proceeding with the
installation. Additionally, a new error type has been added to
specifically handle user configuration errors.

Signed-off-by: Tao Liu <tali@redhat.com>