E2E: Add build tags
This makes use of build tags for the e2e tests and related tools. The
goal is to make it easier to compile and develop BMO by relaxing the
requirement for libvirt bindings. They are now only needed when working
on the relevant code.

One exception is golangci-lint. In order for it to go through all code,
we need to include the build tags when running the linters. Developers
that want to run golangci-lint will need the libvirt bindings.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>