Introduce inspection delay for the fixture provisioner
The fixture provisioner did have a delay previously, but it was removed
as part of some work on reducing access to sensitive information. Now
that we are running tests with the fixture provisioner, this has become
an issue, since inspection happens so fast that it is hard to detect.

This commit introduces artificial delay without accessing any extra
information.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>