Adopt during deprovisioning when custom deploy set
This change ensures that custom deploy methods behave comparably to
image deploys in respect of calling the provisioner's Adopt method.

In practice, the ironic provisioner does not take any action during
Adopt() when the state is deprovisioning, but conceptually it is up to
the provisioner to decide that regardless of the deploy mode.

Signed-off-by: Zane Bitter <zbitter@redhat.com>