Don't inspect custom deployed host on exit from externally provisioned
If an externally provisioned host has a provisioning image set and then
the externallyProvisioned flag is removed, we do not begin an inspection
(which would involve rebooting the host), but instead move directly to
the provisioned state.

This change ensures that the same behaviour occurs when the host has a
custom deploy method set, since this should be equivalent to an image
deploy.

Signed-off-by: Zane Bitter <zbitter@redhat.com>