ironic: Don't adopt during deprovisioning
Since c0373bbfa38d3169d1b994cf0f0429396f12e651 we always set the
instance_info in ValidateManagementAccess() whether or not we are
re-registering a fresh node, so the code to check whether the node has
been re-registered or is transitioning from a clean failure is not
accurate; we have always been adopting when there is an image URL
available, even after a cleaning failed and the node has always existed.
For this code to have had its original effect, we would have to limit
the setting of the instance_info to the Enroll state in
ValidateManagementAccess().

However, since d78fb4d6c69da08e6e41c4e951427509d4ae1784 we always move
the state back to available during deprovisioning, so adopting during
deprovisioning is not needed anyway. The Deprovision() function will
always clean the node.

Therefore, never adopt in Adopt() during deprovisioning, allow
Deprovision() to do the cleaning in both cases.

Signed-off-by: Zane Bitter <zbitter@redhat.com>