Refactor and rename ValidateManagementAccess
Validating access is not its primary responsibility: it's main job is to
make sure the Ironic node exists and is correctly configured. For a lack
of a perfect name, I'm calling it Register here.

The renamed method is moved to a new file and split into more manageable
pieces. In the future, we can even unit-test them separately.

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>