E2E: Delete BMHs prior to namespace deletion
Sometimes the fixture tests hit the timeout for namespace deletion.
The BMO logs indicate that BMO is trying to create new objects while the
namespace is terminating. For example HardwareDetails. To avoid this, I
think we can trigger deletion of the BMHs before we delete the
namespace.

We are running a bit close to the 1m deadline on successful runs
in the re-inspection test.
I believe this is explained by an extra reconcile loop when the
hardwaredetails are updated because of the inspection. No other fixture
test is close to this deadline normally.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>