Redfish: correctly handle missing system ID
Ironic supports [1] not providing a system ID when the BMC handles
exactly one system (which is a common case for standalone servers).
On the other hand, we see a high number of cases where the address is
provided as redfish://host/redfish/v1/, which currently fails with a
very confusing JSON parsing errors.

This change updates all Redfish drivers to ignore paths that are empty
or contain only /redfish/v1. Refactor the drivers to DRY.

[1] https://review.opendev.org/c/openstack/ironic/+/707208

Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>