E2E: Override IRONIC_HTTP_URL
This fixes PXE booting when Ironic cannot be reach directly from the
BMHs. Ironic runs in the 172.18.0.x network (inside kind), but the
relevant ports are exposed on the host. This way the VMs can reach
Ironic at 192.168.222.1.

This commit also updates the e2e README.md, which still had the minikube
configuration.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>