Allow configuring different IPA images per architecture
When using multiple architectures, having a way to set the Ironic
"bootloader" (a.k.a EFI partition) accordingly is important, so this
commit adds a new `DEPLOY_BOOTLOADER_URL` variable to set this Ironic
option.

This commit adds a set of new environment variables allowing to specify
different URLs per target CPU architecture for the IPA image:
 - `DEPLOY_KERNEL_URL_<ARCH>`
 - `DEPLOY_RAMDISK_URL_<ARCH>`
 - `DEPLOY_ISO_URL_<ARCH>`
 - `DEPLOY_BOOTLOADER_URL_<ARCH>`

Non suffixed variables are used as defaults, if there is no architecture
specific image(s) defined for the BMH CPU architecture.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>