E2E: Set static IPs through user-data
We have been "cheating" a bit in CI by configuring static IPs for the
VMs through libvirt dhcp. This is not something we can rely on for real
hardware. For that, we would either need to manually configure DHCP or
make use of cloud-init to configure the IPs. This commit attempts to do
the latter.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>