E2E: Change how upgrade tests are cleaned up
We used DeferCleanup in many places, which is nice in the way that it
can clean up exactly the same thing we have applied. However, it becomes
complicated to know what will be called and in what order.
We have run into situations where BMO is cleaned up before the BMHs are
deleted, causing a dead-lock since nothing can remove the finalizers
anymore.

This commit moves the clean up to AfterEach instead. That way we have
complete control over the order at the cost of not knowing exactly which
kustomization was applied.
It also does the patching of the BMH using Eventually now to avoid
errors if the webhook is a bit slow to start.

A future improvement (larger task) would be to make the upgrade test
declarative instead of generated from the config file, similar to
clusterctl upgrade tests for CAPI. That would allow us to clean up the
correct kustomizations and make it easier to tell from the code what is
being tested.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>