E2E: Use IrSO also in upgrade tests
We are already deploying Ironic using IrSO in the other e2e tests.
Let's do the same for the upgrade tests so we have one unified way of
doing things.
This commit also changes some Ironic/IrSO related variables to be more
in sync with the rest of the variable names. A new variable is added to
make IrSO deployment optional (useful in fixture tests).

Signed-off-by: Lennart Jern <lennart.jern@est.tech>