E2E: Improve logs and artifacts in upgrade tests
This commit is an attempt to improve log and artifact collection in
upgrade tests by storing them in separate folders for each test case.
That way we avoid clobbering logs when they have the same file name and
it is easier to find the relevant files for a specific case.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>